Refactor BlogDetails component to support multi-language content
This commit modifies the BlogDetails component to accept a 'language'
prop, allowing the display of content in different languages. The 'img'
src attribute is updated to use a language-specific image. These changes
enable the website to cater to a diverse audience and improve user
engagement by presenting content in their preferred language.

Issue #123